+++ spade-hir-lowering/src/lib.rs
@@ -26,6 +26,7 @@ use hir::ArgumentList;
 use hir::Attribute;
 use hir::Parameter;
 use hir::TypeDeclKind;
+use hir::TypeSpec;
 use hir::WalTrace;
 use local_impl::local_impl;
 use num::ToPrimitive;
@@ -1729,8 +1730,8 @@ impl ExprLocal for Loc<Expression> {
                 turbofish: _,
             } => {
                 let head = ctx.symtab.symtab().unit_by_id(callee);
-                let args =
-                    match_args_with_params(args, &head.inputs, false).map_err(Diagnostic::from)?;
+                let args = match_args_with_params(args, &head.inputs.inner, false)
+                    .map_err(Diagnostic::from)?;
 
                 match (kind, &head.unit_kind.inner) {
                     (CallKind::Function, UnitKind::Function(_))
@@ -1848,7 +1849,7 @@ impl ExprLocal for Loc<Expression> {
     fn handle_call(
         &self,
         name: &Loc<NameID>,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = StatementList::new();
@@ -2096,7 +2097,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         self.handle_clocked_memory(result, args, ctx, false)
@@ -2106,7 +2107,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         self.handle_clocked_memory(result, args, ctx, true)
@@ -2115,7 +2116,7 @@ impl ExprLocal for Loc<Expression> {
     fn handle_clocked_memory(
         &self,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
         has_initial: bool,
     ) -> Result<StatementList> {
@@ -2221,7 +2222,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         // The localimpl macro is a bit stupid
@@ -2253,7 +2254,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2300,7 +2301,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2361,7 +2362,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2418,7 +2419,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2493,7 +2494,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2523,7 +2524,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2573,7 +2574,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2604,7 +2605,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2634,7 +2635,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2662,7 +2663,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2690,7 +2691,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2718,7 +2719,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2760,7 +2761,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2802,7 +2803,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;
@@ -2832,7 +2833,7 @@ impl ExprLocal for Loc<Expression> {
         &self,
         _path: &Loc<NameID>,
         result: StatementList,
-        args: &[Argument],
+        args: &[Argument<Expression, TypeSpec>],
         ctx: &mut Context,
     ) -> Result<StatementList> {
         let mut result = result;