+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [!304][!304] Allow specifying the bit width of integer literals as `512u32` or `123i64`
 - [!307][!307] Allow specifying command line arguments via a json file
 - [!308][!308] Add `/` and `%` for power of 2 operands, as well as `comb_div` and `comb_mod` for all operands
+- [!309][!309] Add named argument turbofishes (`::$<>`)
 
 ### Changed